pyproject.toml
[project]
name = "arian-receipts"
version = "0.2.0"
requires-python = ">=3.11"
description = "receipt-parser microservice for arian"
readme = "readme.md"
authors = [{ name = "xhos" }]
license = { text = "MIT" }

dependencies = [
  "fastapi>=0.111.0",
	"python-multipart>=0.0.20",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "python-json-logger>=3.3.0",
  "google-genai>=1.27.0",
  "opentelemetry-api>=1.24.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
]

[project.optional-dependencies]
# install with: uv add ".[local]"
local = [
  "pytesseract>=0.3.10",
  "opencv-python-headless>=4.10.0.84",
  "Pillow>=11.3.0",
  "numpy>=2.3.2",
  "llama-cpp-python>=0.2.80",
  "huggingface-hub>=0.24.0",
]

[project.scripts]
arian-receipts = "app.app:main"

[tool.ruff.format]
indent-style = "tab"